+++ mlir/examples/dsp/SimpleBlocks/include/toy/Lexer.h
@@ -47,8 +47,6 @@ enum Token : int {
   // primary
   tok_identifier = -5,
   
-  // test for int and double
-  // tok_number = -6,
   tok_int = -6,
   tok_double = -7,
 };
@@ -87,14 +85,7 @@ public:
     return identifierStr;
   }
 
-  // test for int and double
-  /// Return the current number (prereq: getCurToken() == tok_number)
-  //double getValue() {
-    //assert(curTok == tok_number);
-    //return numVal;
-  //}
-
-  int getIntValue() {
+  int64_t getIntValue() {
       assert(curTok == tok_int);
       return numInt;
   }
@@ -163,18 +154,6 @@ private:
       return tok_identifier;
     }
 
-    // test for int and double
-    // Number: [0-9.]+
-    //if (isdigit(lastChar) || lastChar == '.') {
-      //std::string numStr;
-      //do {
-        //numStr += lastChar;
-        //lastChar = Token(getNextChar());
-      //} while (isdigit(lastChar) || lastChar == '.');
-//
-      //numVal = strtod(numStr.c_str(), nullptr);
-      //return tok_number;
-    //}
 
     if(isdigit(lastChar)) {
         std::string numStr;
@@ -227,10 +206,7 @@ private:
   /// If the current Token is an identifier, this string contains the value.
   std::string identifierStr;
 
-  // test for int and double
-  /// If the current Token is a number, this contains the value.
-  //double numVal = 0;
-  int numInt = 0;
+  int64_t numInt = 0;
   double numDouble = 0;
 
   /// The last value returned by getNextChar(). We need to keep it around as we
